refactor(memory): switch Memory Policy form to tabs
Closes the TODO in frontend/docs/MEMORY_POLICY_CAPTURE_DELEGATION.md. The
form had grown to ~19 fields across 6 cards; groups them into 4 tabs —
Policy / Capture / Retrieval / Guardrails (Write Permissions + Knowledge
Projection + Lifecycle) — reusing the tabConfig/tabFieldMapping/
createFormSubmitHandler pattern from KnowledgeSourceFormPage.tsx.

Each card moved into its own component under
frontend/src/components/memory/tabs/, with the shared SwitchField helper
extracted out of the old MemoryPolicyFields.tsx (deleted) into its own file.
No field logic, validation, or copy changed — pure structural move.

Live-verified against the memory-policy-test bench: all four tabs render
correct fields, tab switching updates the URL hash, Save round-trips
(PUT /api/resource/Memory Policy/... -> 200).
